+++ mmsegmentation/configs/sebnet/sebnet_baseline-p-sbd-head-conditioned_1xb6_cityscapes.py
@@ -54,7 +54,7 @@ model = dict(
                 min_kept=131072,
                 class_weight=class_weight,
                 loss_weight=1.0,
-                loss_name='loss_ce'),
+                loss_name='loss_sem'),
             dict(
                 type='OhemCrossEntropy',
                 thres=0.9,
