+++ mmsegmentation/main.sh
@@ -9,7 +9,7 @@ export CUDA_VISIBLE_DEVICES=3
 #python tools/train.py configs/sebnet/sebnet_baseline-head_1xb8_cityscapes_ohem_scratch.py
 #python tools/train.py configs/sebnet/sebnet_baseline-head_1xb8_cityscapes_ohem.py
 
-python tools/train.py configs/sebnet/sebnet_baseline-d-multilabel-head_1xb6_cityscapes.py
+#python tools/train.py configs/sebnet/sebnet_baseline-d-multilabel-head_1xb6_cityscapes.py
 #python tools/train.py configs/sebnet/sebnet_baseline-casenet-head_1xb6_cityscapes.py
 #python tools/train.py configs/sebnet/sebnet_baseline-bem-head_1xb6_cityscapes.py
 #python tools/train.py configs/sebnet/sebnet_baseline-dff-head_1xb6_cityscapes.py
@@ -24,3 +24,13 @@ python tools/train.py configs/sebnet/sebnet_baseline-d-multilabel-head_1xb6_city
 #python tools/train.py configs/sebnet/sebnet_baseline-casenet-head-earlier-layers_1xb6_cityscapes.py
 #python tools/train.py configs/sebnet/sebnet_baseline-d-multilabel-head-earlier-layers_1xb6_cityscapes.py
 #python tools/train.py configs/sebnet/sebnet_baseline-dff-head-earlier-layers_1xb6_cityscapes.py
+
+#python tools/train.py configs/sebnet/sebnet_baseline-p-sbd-head-conditioned_1xb6_cityscapes.py
+#python tools/train.py configs/sebnet/sebnet_baseline-p-sbd-head-fused_1xb6_cityscapes.py
+#python tools/train.py configs/sebnet/sebnet_baseline-p-sbd-bas-head-conditioned_1xb6_cityscapes.py
+python tools/train.py configs/sebnet/sebnet_baseline-p-sbd-bas-head-fused_1xb6_cityscapes.py
+
+#python tools/train.py configs/sebnet/sebnet_baseline-d-sbd-head_1xb6_cityscapes.py
+#python tools/train.py configs/sebnet/sebnet_baseline-p-d-sbd-head_1xb6_cityscapes.py
+#python tools/train.py configs/sebnet/sebnet_baseline-p-d-sbd-bas-head_1xb6_cityscapes.py
+#python tools/train.py configs/sebnet/sebnet_baseline-p-d-sbd-basWithSBD-head_1xb6_cityscapes.py
